pyproject.toml
[tool.poetry]
name = "django-rest-cheriaa43"
version = "0.1.0"
description = "Django ReST API and Server"
authors = ["Cheria Artis"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.11.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
